refactor: execute migration
