+++ backend/runner/taskrun/schema_update_ghost_cutover_executor.go
@@ -151,7 +151,7 @@ func cutover(ctx context.Context, taskContext context.Context, stores *store.Sto
 		return true, nil, err
 	}
 	defer driver.Close(ctx)
-	migrationID, _, err := utils.ExecuteMigrationWithFunc(ctx, ctx, stores, taskRunUID, driver, mi, statement, &sheetID, execFunc, db.ExecuteOptions{})
+	migrationID, _, err := executeMigrationWithFunc(ctx, ctx, stores, taskRunUID, driver, mi, statement, &sheetID, execFunc, db.ExecuteOptions{})
 	if err != nil {
 		return true, nil, err
 	}
